Miscellaneous fixes.
* download.c: Make various implicit casts explicit.
* download.c: Write title version to opts.version from TMD to prevent
  titleid/0/contentid if called without -V.
* download.c: Fix memory leak in write_file_from_memory when run with
  -m and downloading a title that has no cetk.
* download.c: Remove spurious unreachable return after calls to oom(),
  which terminates the program anyway.
* util.c/README.md: Consistently document -l and that the previous
  default behavior now requires -l.
* util.c: Add -l to usage.
* util.c: Rename __mkdir to MAKEDIR. The C standard reserves
  identifiers that begin with two underscores.
* util.c: Remove unused variable in util_realloc_and_append_fmt.
* util.c: Add -h and -v, present in the usage and README.md, but
  not in the getopt argument list.

+++ crypto.c
@@ -12,7 +12,7 @@ static void rotl(byte *buf)
 	do {
 		set_carry = (buf[i - 1] & 0x80);
 
-		buf[i - 1] = (buf[i - 1] << 1) | carry;
+		buf[i - 1] = (byte)(buf[i - 1] << 1) | carry;
 
 		if (set_carry)
 			carry = 1;
